+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/mediacodec/MediaCodecInfo.java
@@ -34,6 +34,7 @@ import static androidx.media3.exoplayer.mediacodec.MediaCodecPerformancePointCov
 import static androidx.media3.exoplayer.mediacodec.MediaCodecUtil.createCodecProfileLevel;
 import static com.google.common.base.Preconditions.checkNotNull;
 
+import android.content.Context;
 import android.graphics.Point;
 import android.media.MediaCodec;
 import android.media.MediaCodecInfo.AudioCapabilities;
@@ -263,16 +264,17 @@ public final class MediaCodecInfo {
    * Returns whether the decoder may support decoding the given {@code format} both functionally and
    * performantly.
    *
+   * @param context A context.
    * @param format The input media format.
    * @return Whether the decoder may support decoding the given {@code format}.
-   * @throws MediaCodecUtil.DecoderQueryException Thrown if an error occurs while querying decoders.
    */
-  public boolean isFormatSupported(Format format) throws MediaCodecUtil.DecoderQueryException {
+  public boolean isFormatSupported(Context context, Format format) {
     if (!isSampleMimeTypeSupported(format)) {
       return false;
     }
 
-    if (!isCodecProfileAndLevelSupported(format, /* checkPerformanceCapabilities= */ true)) {
+    if (!isCodecProfileAndLevelSupported(
+        context, format, /* checkPerformanceCapabilities= */ true)) {
       return false;
     }
 
@@ -296,12 +298,14 @@ public final class MediaCodecInfo {
   /**
    * Returns whether the decoder may functionally support decoding the given {@code format}.
    *
+   * @param context A context.
    * @param format The input media format.
    * @return Whether the decoder may functionally support decoding the given {@code format}.
    */
-  public boolean isFormatFunctionallySupported(Format format) {
+  public boolean isFormatFunctionallySupported(Context context, Format format) {
     return isSampleMimeTypeSupported(format)
-        && isCodecProfileAndLevelSupported(format, /* checkPerformanceCapabilities= */ false)
+        && isCodecProfileAndLevelSupported(
+            context, format, /* checkPerformanceCapabilities= */ false)
         && isCompressedAudioBitDepthSupported(format);
   }
 
@@ -311,7 +315,7 @@ public final class MediaCodecInfo {
   }
 
   private boolean isCodecProfileAndLevelSupported(
-      Format format, boolean checkPerformanceCapabilities) {
+      Context context, Format format, boolean checkPerformanceCapabilities) {
     Pair<Integer, Integer> codecProfileAndLevel =
         CodecSpecificDataUtil.getCodecProfileAndLevel(format);
     if (format.sampleMimeType != null && format.sampleMimeType.equals(MimeTypes.VIDEO_MV_HEVC)) {
@@ -370,7 +374,7 @@ public final class MediaCodecInfo {
     if (mimeType.equals(MimeTypes.AUDIO_AC4) && profileLevels.length == 0) {
       // Some older devices don't report profile levels for AC-4. Estimate them using other data
       // in the codec capabilities.
-      profileLevels = estimateLegacyAc4ProfileLevels(capabilities);
+      profileLevels = estimateLegacyAc4ProfileLevels(context, capabilities);
     }
     if (SDK_INT == 23 && MimeTypes.VIDEO_VP9.equals(mimeType) && profileLevels.length == 0) {
       // Some older devices don't report profile levels for VP9. Estimate them using other data in
@@ -875,7 +879,7 @@ public final class MediaCodecInfo {
    * @return The estimated {@link CodecProfileLevel CodecProfileLevels} for the decoder.
    */
   private static CodecProfileLevel[] estimateLegacyAc4ProfileLevels(
-      @Nullable CodecCapabilities capabilities) {
+      Context context, @Nullable CodecCapabilities capabilities) {
     int maxInChannelCount = 2;
     if (capabilities != null) {
       @Nullable AudioCapabilities audioCapabilities = capabilities.getAudioCapabilities();
@@ -887,6 +891,15 @@ public final class MediaCodecInfo {
     int level = CodecProfileLevel.AC4Level3;
     if (maxInChannelCount > 18) { // AC-4 Level 3 stream is up to 17.1 channel
       level = CodecProfileLevel.AC4Level4;
+    }
+
+    // Automotive platform MediaCodec AC-4 decoders are not expected to support AC4Profile22 and
+    // deprecated profiles AC4Profile00, AC4Profile10 and AC4Profile11. See
+    // https://professionalsupport.dolby.com/s/article/Support-Dolby-Atmos-in-Android-Automotive-OS-media-apps?language=en_US
+    if (Util.isAutomotive(context)) {
+      return new CodecProfileLevel[] {
+        createCodecProfileLevel(CodecProfileLevel.AC4Profile21, level)
+      };
     }
 
     return new CodecProfileLevel[] {